Fullscreen toggle, FPS monitor, HSVA control, add assets/config.json
The config file is loaded on program start and saved on exit. Closes #5.

The visualisation will start in fullscreen with mouse hidden if the
toggle is active. When inactive, windows will no longer overlap when
opened.

An fps.rs module was added for monitoring visualisation FPS. This same
type could be used for monitoring serial-over-USB rates once added.
Closes #3.

HSVA is passed to the fragment shader as uniform data.

Various GPU buffer re-use optimisations have been applied.