Add new `Decay` frame buffer and render pass. Add `sustain` param.
Previously, the phosphor decay affect was achieved solely by applying a
<1.0 alpha value to the output drawn to the window's swapchain image. In
other words, on each frame we would load the previous state of the
swapchain image and then overwrite it by an amount specified by the
`alpha` value.

This PR removes this old approach in favour of using a dedicated `Decay`
buffer. This allows for tuning the "sustain" level of the decay.

Rather than using a single render pass that draws directly to the
swapchain image, we now use two:

1. Render the decay image. The decay is achieved using the same approach
   as was applied previously, i.e. using an alpha value to specify the
   "amount" that the previous state of the buffer should be overwritten.
2. Render the swapchain image (window image). Achieved by selecting the
   max between the `decay * sustain` and the currently presented
   character if there is one.

The new `sustain` parameter is applied to the decay in the *second*
renderpass, right before selecting the max.

The new decay buffer uses a 32-bit floating point pixel format to ensure
a smooth perceived decay.

We now also recreate the graphics pipeline if the window is resized to
ensure the decay buffer size always precisely matches the swapchain
image size. This allows for pixel-perfect accuracy and avoids strange
artifacts that can occur around the edges of the characters when
sampling from a texture of a slightly different size.